# Faster tokenization by converting byte arrays to strings