… go/vt/sqlparser/analyzer_test.go
… go/vt/sqlparser/ast_test.go
… go/vt/sqlparser/parse_next_test.go
… go/vt/sqlparser/parse_test.go
… go/vt/sqlparser/random_expr.go
… go/vt/sqlparser/sql.go
… go/vt/sqlparser/token.go
… go/vt/sqlparser/token_test.go
… go/vt/vtgate/engine/memory_sort_test.go
… go/vt/vtgate/planbuilder/expr_test.go
… go/vt/vtgate/planbuilder/select.go
… go/vt/vttablet/tabletmanager/vreplication/table_plan_builder.go
… go/vt/vttablet/tabletserver/planbuilder/plan.go
… go/vt/vttablet/tabletserver/vstreamer/uvstreamer.go
… go/vt/vttablet/vexec/vexec.go
… go/vt/wrangler/materializer.go
… go/vt/wrangler/stream_migrater.go
… go/vt/wrangler/vexec_plan.go